backend/requirements.txt
boto3
PyJWT==2.8.0
